Concurrency-safety!
The implementation was not concurrency-safe, since the working
directory of the process was modified for each request. So it
would behave incorrectly when processing multiple requests in
parallel.

This situation occurs more frequently than previously, since we
are now listening on several branches of the flourish repository,
and so a push to both the master and live branches will trigger
simultaneous requests.